Split command logic from object file
